2. Create Your Magical Tables

+++ task_2.sql
@@ -9,7 +9,7 @@ CREATE TABLE Books (
     book_id INT PRIMARY KEY,
     title VARCHAR(130),
     author_id INT,
-    price FLOAT,
+    price DOUBLE,
     publication_date DATE,
     FOREIGN KEY(author_id) REFERENCES Authors(author_id)
 );
@@ -36,7 +36,7 @@ CREATE TABLE Order_Details(
     orderdetailid INT PRIMARY KEY,
     order_id INT,
     book_id INT,
-    quantity FLOAT,
+    quantity DOUBLE,
     FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     FOREIGN KEY (book_id) REFERENCES Books(book_id)
 );
